bleCollector/requirements.txt
python-dotenv
flask
flask-mqtt
line-bot-sdk
pyngrok
pymongo
